Added jsdocs for the common API modules
